Add Character class, which extends Entity, and extend it by Player and NonPlayableCharacter
